Fixes date reset.
When note has reset instructions, all versions are destroyed before
save, date updated is set to date created (on cloud), and no new version
is created on save. When always_reset_on_create is true, date updated is
set to date created (on cloud) when a note is first created.