Fix channels caching
Removed all uncached calls to channels. The problem it was trying to
solve was that when a user edits their page, they could end up thinking
nothing has changed because they are viewing cached pages. Rails is set
not to put a cache header if the channel has been edited more recently
than ten minutes ago but if the user has already seen some pages, the
browser would not request them for a while. Howver, setting pjax cache:
false is undesirable since it puts in a cache_buste rin the url. A more
sophisticated approach is required. Rails can check when user last
signed in and 'uncache' all her channels for ten minutes. But this would
not affect the first page she sees, or any other pages the user had seen
in the ten minutes before she logged in. (She will always see at least
one page, the one that makes the AJAX call to the user menu.)